pyproject.toml
[tool.poetry]
name = "repl_python3_LikelyLightyellowMining"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
instaloader = "^4.7.1"
discord = "^1.0.1"
python-dotenv = "^0.17.1"
"discord.py" = "^1.7.2"
flask = "^2.0.0"
emoji = "^1.2.0"
jj = "^2.0.1"
pytz = "^2021.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
